Rename geom to math
