+++ src/lib.rs
@@ -10,7 +10,7 @@ mod main_loop;
 pub(crate) mod util;
 
 pub mod draw;
-pub mod geom;
+pub mod math;
 
 // Re-export dependencies that occur in our public API.
 pub use golem;
@@ -21,6 +21,6 @@ pub use canvas::Canvas;
 pub use color::{Color3, Color4};
 pub use draw::{Batch, Font, TextBatch, Texture};
 pub use error::Error;
-pub use geom::{Camera, Rect, Screen};
+pub use math::{Camera, Rect, Screen};
 pub use input::{Event, InputState, Key};
 pub use main_loop::main_loop;
